sdk/debug/lost-context-simulator-test.html

<!-- this file is auto-generated. DO NOT EDIT.

/*
** Copyright (c) 2012 The Khronos Group Inc.
**
** Permission is hereby granted, free of charge, to any person obtaining a
** copy of this software and/or associated documentation files (the
** "Materials"), to deal in the Materials without restriction, including
** without limitation the rights to use, copy, modify, merge, publish,
** distribute, sublicense, and/or sell copies of the Materials, and to
** permit persons to whom the Materials are furnished to do so, subject to
** the following conditions:
**
** The above copyright notice and this permission notice shall be included
** in all copies or substantial portions of the Materials.
**
** THE MATERIALS ARE PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
** EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
** MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.
** IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY
** CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT,
** TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
** MATERIALS OR THE USE OR OTHER DEALINGS IN THE MATERIALS.
*/

-->
<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN"
  "http://www.w3.org/TR/html4/loose.dtd">
<html>
  <head>
    <title>WebGL Lost Context Wrapper Test</title>
    <link rel="stylesheet" href="../tests/resources/js-test-style.css"/>
    <script src="../tests/resources/js-test-pre.js"></script>
    <script src="../tests/conformance/resources/webgl-test.js"> </script>
    <script src="webgl-debug.js"> </script>
</head>
<body>
<canvas id="example" width="2" height="2" style="width: 40px; height: 40px;"></canvas>
<div id="description"></div>
<div id="console"></div>
<script>
debug("Tests the context lost simulator.");
debug("");
var c = document.getElementById("example");
c = WebGLDebugUtils.makeLostContextSimulatingCanvas(c);
c.addEventListener('webglcontextlost', handleContextLost, false);
c.addEventListener('webglcontextrestored', afterContextRestored, false);
var gl = create3DContext(c);

c.loseContextInNCalls(5);

assertMsg(gl, "Got 3d context.");

var tex = gl.createTexture();
gl.bindTexture(gl.TEXTURE_2D, tex);
gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_WRAP_S, gl.CLAMP_TO_EDGE);
glErrorShouldBe(gl, gl.NO_ERROR, "no errors from setup");
glErrorShouldBe(gl, gl.CONTEXT_LOST_WEBGL, "error should be lost context");
shouldBeNull("gl.createBuffer()");

function handleContextLost(e) {
  e.preventDefault();
}

function afterContextRestored(e) {
  glErrorShouldBe(gl, gl.NO_ERROR, "no errors after context restored");
  gl.bindTexture(gl.TEXTURE_2D, tex);
  glErrorShouldBe(gl, gl.NO_ERROR, "no errors binding a resource from lost context");
  gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_WRAP_S, gl.CLAMP_TO_EDGE);
  glErrorShouldBe(gl, gl.INVALID_OPERATION, "invalid operation using a resource from lost context");
  tex = gl.createTexture();
  gl.bindTexture(gl.TEXTURE_2D, tex);
  gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_WRAP_S, gl.CLAMP_TO_EDGE);
  glErrorShouldBe(gl, gl.NO_ERROR, "okay to use new context");

  successfullyParsed = true;
  var epilogue = document.createElement("script");
  epilogue.onload = finish;
  epilogue.src = "../tests/resources/js-test-post.js";
  document.body.appendChild(epilogue);
}

function finish() {
    if (window.nonKhronosFrameworkNotifyDone) {
        window.nonKhronosFrameworkNotifyDone();
    }
}

</script>
</body>
</html>
